Added Atomic Calculator stand alone functions python file
